float to int

+++ app.py
@@ -42,7 +42,7 @@ def user_input_features():
     property_tax = st.sidebar.slider('propTax', float(0.5), float(1), float(0.8))
     living_area = st.sidebar.slider('livingArea', int(100), int(8000), int(2500))
     lot_size_sqft = st.sidebar.slider('lot_size_sqf', int(0), int(10000), int(5000))
-    school_rank = st.sidebar.slider('schlRnk', int(0), float(10), float(5))
+    school_rank = st.sidebar.slider('schlRnk', int(0), int(10), int(5))
     hoa = st.sidebar.slider('hoa', float(0), float(2000), float(300))
     has_fireplace = st.sidebar.slider('hasFireplace', int(0), int(1), int(1))
     garage_spaces = st.sidebar.slider('garageSpaces', int(0), int(10), int(2))
